+++ src/mir/repres/unsupported/HEALPixNested.cc
@@ -17,6 +17,7 @@
 #include "eckit/log/JSON.h"
 
 #include "mir/iterator/UnstructuredIterator.h"
+#include "mir/reorder/HEALPix.h"
 #include "mir/util/Exceptions.h"
 #include "mir/util/Grib.h"
 #include "mir/util/GridBox.h"
@@ -55,14 +56,14 @@ void HEALPixNested::print(std::ostream& out) const {
 
 
 std::vector<util::GridBox> HEALPixNested::gridBoxes() const {
-    const proxy::HEALPix::Reorder reorder(static_cast<int>(ring_.Nside()));
-    const auto N = numberOfPoints();
+    const auto N       = numberOfPoints();
+    const auto reorder = reorder::HEALPixRingToNested(N).reorder();
 
     std::vector<util::GridBox> boxes(N);
 
     int i = 0;
     for (const auto& box : ring().gridBoxes()) {
-        auto j      = reorder.ring_to_nest(i++);
+        auto j      = reorder.at(i++);
         boxes.at(j) = box;
     }
     ASSERT(i == N);
@@ -79,15 +80,15 @@ std::vector<util::GridBox> HEALPixNested::gridBoxes() const {
 
 Iterator* HEALPixNested::iterator() const {
     if (longitudes_.empty()) {
-        const proxy::HEALPix::Reorder reorder(static_cast<int>(ring_.Nside()));
-        const auto N = numberOfPoints();
+        const auto N       = numberOfPoints();
+        const auto reorder = reorder::HEALPixRingToNested(N).reorder();
 
         longitudes_.resize(N);
         latitudes_.resize(N);
 
         int i = 0;
         for (const auto& point : ring().atlasGrid().lonlat()) {
-            auto j            = reorder.ring_to_nest(i++);
+            auto j            = reorder.at(i++);
             longitudes_.at(j) = point.lon();
             latitudes_.at(j)  = point.lat();
         }
